requirements.txt
python-dotenv
python-telegram-bot
pandas
pandas_ta
watchdog
aiofiles
pydantic
colorama
